debug

+++ headers/myhashstoragefile.h
@@ -11,7 +11,7 @@ typedef struct node{
 	char* abs_path;
 	char* filename;
 	void* content;
-	unsigned long dim_bytes;
+	size_t dim_bytes;
 	int modified_flag;
 	int opened_flag;
 	int o_create_flag;
@@ -34,16 +34,16 @@ typedef struct temp_list{
 
 typedef struct _list{
 	int size;
-	unsigned long dim_bytes;
+	size_t dim_bytes;
 	file_t *head;
 	file_t *tail;
 }list;
 
 typedef struct _hash{
 	int len;
-	unsigned long memory_used;
-	unsigned long memory_used_from_modified_files;
-    unsigned long memory_capacity;
+	size_t memory_used;
+	size_t memory_used_from_modified_files;
+    size_t memory_capacity;
 	int n_file;
 	int n_files_free;
 	int n_file_modified;
@@ -51,7 +51,7 @@ typedef struct _hash{
     int max_size_cache;
 
 	int stat_max_n_file;
-	unsigned long stat_dim_file;
+	size_t stat_dim_file;
 	int stat_n_replacing_algoritm;
 	
 	list *cell;
